Add comment motivating use of lazy in publisher (#1527)
https://github.com/zio/zio-kafka/pull/1437 introduced a clever way to
retry records publication after specific errors (more details in the
referred PR).
There is a place of `lazy val` usage in that implementation, which may
raise questions of why we need it, from the folks unfamiliar with the
implementation.
To add a bit more clarity, this PR adds a comment with the reasoning for
using `lazy val`.