Add implicit trace parameter
Add implicit trace parameter to all public and private methods that return a ZIO, as recommended by the [guidlines for library author](https://zio.dev/guides/migrate/zio-2.x-migration-guide#guidelines-for-library-authors).

Warning: this change is not binary, nor source compatible. Source compatibility is only broken when a trait of zio-kafka is implemented by the user (for example the `Serde` trait).