Limit size of consumer batches to max.poll.records
During shutdown, when rebalanceSafeCommits is enabled, we wait until the consumer has processed all records that were pulled from  the partition queue. Before this change, the consumer would pull all available (pre-fetched) records. We now put an upper limit on the amount of records that can be pulled. This decreases the time we need to wait, while it doesn't hurt throughput too much.

Note: this does mean we zio-kafka can no longer guarantee that all records (for a given partition) fetched together from the broker are grouped into one chunk.

This change also extracts collecting the runloop configuration to a separate class. This has the additional advantage that it can be tested better.