Disallow CooperativeStickyAssignor for transactional, make RangeAssignor the default for rebalance-safe-commits
**Context:** most partition assignors revoke all partitions during a rebalance, even when a partition is re-assign to the same consumer after the rebalance. The CooperativeStickyAssignor however, has the option to not revoke a partition in case it will be assigned to the same consumer.

When rebalance-safe-commits is enabled, the zio-kafka consumer awaits processing completion for every partition that gets revoked during a rebalance. Partitions that are not revoked in the rebalance, will continue to process records. _After_ the rebalance, the consumer might still be processing records that were received from _before_ the rebalance. In cases where the epoch of the consumed records is important, this causes problems.

One such case is when the consumer is coupled to a transactional producer. The brokers do not allow committing a transaction with offsets received in a previous epoch. To prevent this situation, zio-kafka now validates that the CooperativeStickyAssignor is not configured for transactional use cases.

Since zio-kafka does not know upfront whether a consumer will be coupled to a transactional producer, and because the default partition assignor is `[RangeAssignor, CooperativeStickyAssignor]` (the consumer switches to the second when all consumers support it), and because it is not possible to receive new records during the rebalance anyways (this is a limitation of the underlying kafka consumer, see also KIP-983), we now make `[RangeAssignor]` the default when rebalance-safe-commits is enabled.

The custom `RebalanceSafeCommitsRequired` (a subclass of IllegalArgumentException) has been removed. We now use IllegalArgumentException directly.

Also:
- Test transactional producing with all standard partition assignors (except the CooperativeStickyAssignor).
- Make the transactional producing test fail immediately when constructing the consumer fails.
- Mark some methods on ConsumerSettings as an internal API.
- More ConsumerSettings tests.
- Make the error message for invalid settings a bit nicer.