Await completion of some admin operations (#1611)
Zio-kafka currently has a fixed delay as a workaround for
[KAFKA-18818](https://issues.apache.org/jira/browse/KAFKA-18818).
However, as we learned recently, the admin operations are eventual
consistent; a delay does not guarantee anything.

We replace the workaround by a proper fix: repeatedly read from the
broker until we see that the operation has finished. To prevent a
complete halt on weird edge cases we wait a limited time (5s).

In this PR we only address create and delete topic operations. If the
approach works, we can extend it to other operations.

Also:
- Extend the workaround to all admin operations that change something.
- Reduce test parallelism to reduce pressure on kafka broker.